requirements.txt
azure-functions
azure-storage-file-share
azure-storage-blob
azure-ai-formrecognizer
requests